LayoutTests/http/tests/w3c/html/semantics/document-metadata/the-title-element/title.text-02-expected.txt


FAIL COMMENT assert_equals: expected "TEXT" but got "TEXTELEMENT"
